+++ src/apify/storages/dataset.py
@@ -461,6 +461,7 @@ class Dataset(BaseStorage):
         id: Optional[str] = None,
         name: Optional[str] = None,
         force_cloud: bool = False,
+        custom_client: Optional[ApifyClientAsync] = None,
         config: Optional[Configuration] = None,
     ) -> 'Dataset':
         """Open a dataset.
@@ -478,9 +479,10 @@ class Dataset(BaseStorage):
                 If the dataset with the given name does not exist, it is created.
             force_cloud (bool, optional): If set to True, it will open a dataset on the Apify Platform even when running the actor locally.
                 Defaults to False.
+            custom_client (ApifyClientAsync, optional): A custom instance of the `ApifyClientAsync` class.
             config (Configuration, optional): A `Configuration` instance, uses global configuration if omitted.
 
         Returns:
             Dataset: An instance of the `Dataset` class for the given ID or name.
         """
-        return await super().open(id=id, name=name, force_cloud=force_cloud, config=config)
+        return await super().open(id=id, name=name, force_cloud=force_cloud, custom_client=custom_client, config=config)